feat(compress): move to the compress package
