--- compressor/module.go
@@ -1,11 +1,0 @@
-package compressor
-
-import (
-	"go.uber.org/fx"
-)
-
-// Module for fx.
-var Module = fx.Options(
-	fx.Provide(NewSnappy),
-	fx.Provide(NewMap),
-)
